% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trmd}
\alias{plot.trmd}
\title{plot() Method for Objects of Class 'trmd'}
\usage{
\method{plot}{trmd}(object, ...)
}
\arguments{
\item{object}{Objects of Class 'trmd'}

\item{...}{additional arguments}
}
\value{
a histogram plots.
}
\description{
plot() Method for Objects of Class 'trmd'
}
