% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trme}
\alias{plot.trme}
\title{plot() Method for Objects of Class 'trme'}
\usage{
\method{plot}{trme}(object, ...)
}
\arguments{
\item{object}{Objects of Class 'trme'}

\item{...}{additional arguments}
}
\value{
a histogram plots.
}
\description{
plot() Method for Objects of Class 'trme'
}
