% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.trmd}
\alias{print.trmd}
\title{print() Method for Objects of Class 'trmd'}
\usage{
\method{print}{trmd}(object)
}
\arguments{
\item{object}{Objects of Class 'trmd'.}
}
\value{
print the summarized results.
}
\description{
print() Method for Objects of Class 'trmd'
}
