% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.trme}
\alias{print.trme}
\title{print() Method for Objects of Class 'trme'}
\usage{
\method{print}{trme}(object)
}
\arguments{
\item{object}{Objects of Class 'trme'.}
}
\value{
print the summarized results.
}
\description{
print() Method for Objects of Class 'trme'
}
