% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.trme}
\alias{summary.trme}
\title{Summary() Method for Objects of Class 'trme'}
\usage{
\method{summary}{trme}(object)
}
\arguments{
\item{object}{Objects of Class 'trme'}
}
\value{
a data frame containing summarized estimates of parameters.
}
\description{
Summary() Method for Objects of Class 'trme'
}
