% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trme.R
\name{trme}
\alias{trme}
\title{Title Triple Robust Estimation for Missing Exposure}
\usage{
trme(
  covs,
  Y,
  A,
  df_mar,
  ps_model = T,
  imp_model = T,
  quan_value = 0.99,
  method = c("new", "ee")
)
}
\arguments{
\item{covs}{\code{\link{character}}  the vector name of all confounders which
will affect both exposure and outcome.  Any interaction term or non-linear
form can be added after creating those variables based on the data. Factors
with two or multiple levels are allowed and will be converted as numerical
values.}

\item{Y}{\code{\link{character}} the name of the outcome variable. The trme
function currently only works on the binary outcome.}

\item{A}{\code{\link{character}} the name of binary treatment or exposure
variable.}

\item{df_mar}{a data frame contains the variables in the models.}

\item{ps_model}{a logical value either TRUE or FALSE. ps_model refers to the
treatment and imputation model, respectively. If the model=FALSE, the Bayes
rule will be applied to estimate it.}

\item{imp_model}{a logical value either TRUE or FALSE. imp_model refers to
the treatment and imputation model, respectively. If the model=FALSE, the
Bayes rule will be applied to estimate it.}

\item{quan_value}{\code{\link{numeric}} shrinkage value. By default, we
shrink weights larger than 99\% quantile into exact 99\% quantile to avoid
extreme weights.}

\item{method}{the method to be used. either proposed "new" TR estimator or "ee" complex TR estimator.}
}
\value{
A list ...
\itemize{
    \item{Estimate}{estimated causal effect (log odds ratio) of exposure on the outcome. }
\item{Robust.SE}{estimated robust standard errors used for inference.}
\item{p.value}{p values for two-sided Wald test.}\item{95\% CI}{95\% two-sided confidence interval.In addition, some fitted values and plots are also provided including:}
\item{vcov}{variance-covariance matrix among exposure and covariates.}
\item{fit_ps_all}{predicted propensity score values for all subjects used to adjust for the confounding issue.}
\item{fit_weightmiss}{fitted inverse weights of missingness used to adjust for the missing issue.}
\item{hist_ps_control, hist_ps_trt}{histogram for predicted propensity score between control and treatment groups.}
}
}
\description{
Estimate the causal effect of the exposure on the outcome when the exposure is MAR. Adjust for both missing and confounding issues via simplified estimating equations with triple robust (TR) properties. Provide robust standard errors for inference purposes.
}
\details{
The methods currently only work for the situation of missing
  exposure and require binary outcome and exposure variables. If either the
  covariates or outcome is also missing, first impute missing values based on
  \code{\link{mice}}. The basic assumption is the exposure variable is
  missing at random (MAR), which means that given all observed covariates and
  outcomes, the missingness (or missing indicator) is independent of the
  missing value itself. If the exposure is missing not at random (MNAR), try
  another method instead. \code{method="new"} is a simplified estimating
  equation for the triple robust (TR) estimator (recommended), which speeds
  up the computation process and avoids some effects of extreme weights in
  the finite samples, but still keeps the same TR properties as the complex
  form. For more details, please review the reference paper.
  \code{method="ee"} is a complex estimating equation for TR estimator, which
  may be influenced by extreme weights. Both two TR estimators have the same
  asymptotical consistency when the sample size is large. To achieve
  consistency, both TR estimators require at least ``one of two models"
  condition, which means if the missingness model is correct, we require
  either the treatment or outcome model to be correct; If the missingness
  model is wrong, but the outcome model is correct, we require either the
  imputation or the treatment model to be correct. The asymptotical standard
  errors may not be the same. However, both TR estimators utilize robust
  standard error (RSE) based on the sandwich formula, which is quite robust
  to the model's misspecification.
}
\references{
Yuliang Shi, Yeying Zhu, Joel Dubin. \emph{Causal Inference on Missing Exposure via Triple Robust Estimator}. Statistics in Medicine. Submitted (11/2022).
}
