% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.trmd}
\alias{summary.trmd}
\title{summary() Method for Objects of Class 'trmd'}
\usage{
\method{summary}{trmd}(object)
}
\arguments{
\item{object}{Objects of Class 'trmd'}
}
\value{
a list containing estimates of parameters.
}
\description{
summary() Method for Objects of Class 'trmd'
}
