% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trme.R
\name{trme}
\alias{trme}
\title{Triple Robust Estimation for Missing Exposure}
\usage{
trme(
  covs,
  Y,
  A,
  data,
  imp_model = T,
  shrink_rate = 1,
  ci_alpha = 0.95,
  method = c("TR-AIPW", "TR-WEE"),
  bootstrap = T,
  B = 200
)
}
\arguments{
\item{covs}{\code{\link{character}}  the required vector name of all confounders which
will affect both exposure and outcome.  Any interaction term or non-linear
form can be added after creating those variables based on the data.}

\item{Y}{\code{\link{character}} the required name of the outcome variable. The trme
function currently only works on the binary outcome.}

\item{A}{\code{\link{character}} the required name of binary treatment or exposure
variable.}

\item{data}{a required \code{\link{data.frame}} contains the variables in the models.}

\item{imp_model}{a \code{\link{logical}} value either \code{TRUE} or \code{FALSE} for correct or wrong imputation model. If the model=\code{FALSE}, the Bayes rule will be applied to estimate it.}

\item{shrink_rate}{\code{\link{numeric}} shrinkage rate. By default, no shrinkage applies.
In some cases, we can shrink weights larger than 99\% quantile into exact 99\% quantile to avoid
extreme weights.}

\item{ci_alpha}{\code{\link{numeric}} 0\%-100\% percentage of confidence interval. By default,  95\% CI is present.}

\item{method}{\code{\link{character}} the method to be used. either "AIPW" for TR estimator using augmented inverse-probability weighting or "WEE" for TR estimator using weighted estimating equations.}

\item{bootstrap}{a \code{\link{logical}} value either \code{TRUE} or \code{FALSE} for bootstrap method. If \code{TRUE}, provide statistical reference. If \code{FALSE}, only return point estimate.}

\item{B}{a \code{\link{numeric}} value for replication times of bootstrap.}
}
\value{
Use \code{summary()} function to print out a data frame including summarized results.
\itemize{
\item{\code{Estimate: }}{estimated causal effect (odds ratio) of exposure on the outcome. }
\item{\code{95\% CI BSE: }}{95\% two-sided confidence interval using bootstrap standard errors and normal approximation.}
#' \item{\code{95\% CI Per: }}{95\% two-sided confidence interval using bootstrap percentile (recommended).}
\item{\code{p.value: }}{p values for two-sided Wald test. H0: true causal effect (OR) is 1}}

In addition, other fitted values are also saved in the list.
\itemize{
\item{\code{fit_ps: }}{fitted propensity scores for all subjects, which are used to adjust for the confounding issue.}
\item{\code{miss_weights: }}{fitted inverse weights of missingness used to adjust for the missing issue.}
\item{\code{hist_ps_control, hist_ps_trt: } use \code{plot()} function to draw density plots.}
}
}
\description{
Estimate the causal effect of the exposure on the outcome when the exposure is MAR. Adjust for both missing and confounding issues using triple robust (TR) estimator via weighted estimating equations. Provide estimated standard errors for inference purposes.
}
\details{
The methods currently only work for the missing
  exposure with binary outcome and exposure variables. If either the
  covariates or outcome is also missing, please impute missing values via \code{\link{mice}}.

  The basic assumption is that the exposure variable is
  missing at random (MAR), i.e. given all observed covariates and
  outcomes, the missingness (or missing indicator) is independent of the
  missing value itself. If the exposure is missing not at random (MNAR), try
  another method instead.

  \code{method="TR-AIPW"} is for the TR AIPW estimator, which requires "two correct models from three groups (missing/imputation model group, treatment model group, outcome model group)".

  \code{method="TR-WEE"} is a weighted estimating equation for TR estimator (recommended), which avoids some effects of extreme weights in the finite samples, but it still keeps the same TR properties as the complex form. For more details, please review the reference paper.

  Both two TR estimators have the same
  asymptotic consistency when the sample size is large. To achieve
  consistency, both TR estimators require at least **two correct groups from three groups of models**
  , which means if either the missingness or imputation model is correct, we require
  either the treatment or outcome model to be correct;  If both imputation and missingness models are wrong, both treatment and outcome models require to be correct to acheive consistency, in order to apply Bayes approach.

Both TR estimators utilize Bootstrap approach to estimate standard errors and CI which can protect against
  the misspecification of model based on the simulation studies. By default, parallel computing will be applied to speed up computing process based on the operating system.
}
\note{
For more details, please review \href{https://github.com/yuliang-shi/trme}{Yuliang's Github}.
For citation, please cite the package as Yuliang Shi, Yeying Zhu, Joel Dubin. \emph{Causal Inference on Missing Exposure via Robust Estimator}. Biometrical Journal. Submitted.
}
\examples{
########The first example for simulated data##########
require("trme")
set.seed(2021)
n = 1500 #sample size

#####generate some continuous covariates
id = seq(1, n, by = 1)
x1 = rnorm(n)
x2 = rnorm(n)
x3 = rnorm(n)

#generate binary exposure from the PS model. x1,x2,x3 as confounders
a = -0.2 + 0.8 * x1 + 0.9 * x2 + 1 * x3
prob_a = 1 / (1 + exp(-a))
A = rbinom(n, 1, prob_a)

#generate binary outcome from OR model
z = 0.9 + 1 *A + 0.9 * x1 + 0.6 * x2 + 0.5 * x3   #-1.5 is ok
Y = rbinom(n, 1, 1 / (1 + exp(-z)))      #' bernoulli response variable

##df: before remove missing data
df = data.frame(
  id = id,
  Y = Y,
  A = factor(A),
  x1 = x1,
  x2 = x2,
  x3 = x3,
prob_a = prob_a
)

##control the miss rate as 20\\%
r_sim = -1 +0.4 * x1 +0.6 * x2 +0.8 * x3 - 0.9 * Y #add random error
r = rbinom(n, 1, 1 / (1 + exp(-r_sim))) #miss rate

##data: after include missing data
data = cbind(df, "r" = r)
data$A[which(data$r == 1)] = NA

##test in the simulated data
##true causal (odds ratio) is 2.247 in this case
##use WEE as estimate
tr_wee = trme(
  covs = c("x1", "x2", "x3"),
  Y = "Y",
  A = "A",
  data = data,
  imp_model = T,
  shrink_rate = 1,
  ci_alpha=0.95,
  method = "TR-WEE",
  bootstrap=T,
  B=500
)

##print out results and plots
summary(tr_wee)

##draw plots
##plot(tr_wee)

##apply AIPW method
tr_aipw = trme(
  covs = c("x1", "x2", "x3"),
  Y = "Y",
  A = "A",
  data = data,
  imp_model = T,
  shrink_rate = 1,
  ci_alpha=0.95,
  method = "TR-AIPW",
  bootstrap=T,
  B=500
)

##print out results
summary(tr_aipw)

##draw plots
##plot(tr_aipw)


########The second example for real data##########
require("trme")
data(covid19)

##use TR WEE estimator
tr_wee=trme(covs = c("age","sex","diabetes"),Y="Y",A="CVD", data=covid19,
            imp_model=T,shrink_rate = 1,ci_alpha=0.95,
            method="TR-WEE",bootstrap=T,B=200)

##estimate of causal effect and CI
summary(tr_wee)

##use TR AIPW method
tr_aipw=trme(covs = c("age","sex","diabetes"),Y="Y",A="CVD", data=covid19,
           imp_model=T,shrink_rate = 1,ci_alpha=0.95,
           method="TR-AIPW",bootstrap=T,B=200)

##use TR AIPW method
summary(tr_aipw)
}
\references{
Yuliang Shi, Yeying Zhu, Joel Dubin. \emph{Causal Inference on Missing Exposure via Robust Estimator}. Statistics in Medicine. Biometrical Journal. Submitted.

Zhang, Z., Liu, W., Zhang, B., Tang, L., and Zhang, J. (2016). \emph{Causal inference with missing exposure information: Methods and applications to an obstetric study}. Statistical Methods in Medical Research 25, 2053–2066.

Williamson, E. J., Forbes, A., and Wolfe, R. (2012). \emph{Doubly robust estimators of causal exposure effects with missing data in the outcome, exposure or a confounder}. Statistics in medicine 31, 4382–4400.
}
\seealso{
\code{\link{summary.trme}}, \code{\link{print.trme}} for summarized result; \code{\link{plot.trme}} for histograms of fitted propensity score; \code{\link{covid19}} for description of real data set.
}
\author{
Yuliang Shi
}
\keyword{regression,}
\keyword{robust.}
